src/main/resources/requirements.txt
psycopg2-binary
python-slugify
python-dotenv
